Dev/lerplerplerp

+++ nnue/src/preprocess/load_write.rs
@@ -10,13 +10,17 @@ use utils::memory::boxed_zeroed;
 use crate::arch::{NNUEData, QuantNNUEData, RawNNUEData};
 
 pub trait LoadWrite: Sized {
+    /// Load the given type from some file.
+    ///
+    /// # Errors
+    ///     Errors when the file cannot be opened, or does not match the file size.
     fn load_from_file(path: &Path) -> Result<Box<Self>> {
         let mut file = File::open(path)?;
 
         let expected = std::mem::size_of::<Self>();
-        let actual = file.metadata()?.len() as usize;
+        let actual = file.metadata()?.len();
 
-        if expected != actual {
+        if (expected as u64) != actual {
             return Err(Error::raw(
                 ErrorKind::InvalidValue,
                 format!("Error loading {}: Expected {expected} bytes, found {actual} bytes!", std::any::type_name::<Self>()),
@@ -31,6 +35,10 @@ pub trait LoadWrite: Sized {
         }
     }
 
+    /// Write the given type to some file.
+    ///
+    /// # Errors
+    ///     Errors when the file cannot be written to.
     fn write_to_file(&self, path: &Path) -> Result<()> {
         let mut file = File::create(path)?;
         let len = std::mem::size_of::<Self>();
